readyset-data: Fix PG real data type representation
When lowering MIR into dataflow, we are currently represent PG `real`
data types as `DfType::Double`. According to the PG docs [0], `real`
is a 4-byte floating-point value, while our `double` is an 8-byte
floating-point value.

This CL corrects that to represent `real` types as `float`s.

[0] https://www.postgresql.org/docs/current/datatype-numeric.html

Fixes: REA-4193

+++ readyset-data/src/dialect.rs
@@ -96,7 +96,7 @@ impl Dialect {
     pub(crate) fn real_type(&self) -> DfType {
         // TODO: Handle `real_as_float` mode.
         match self.engine {
-            SqlEngine::PostgreSQL => DfType::Double,
+            SqlEngine::PostgreSQL => DfType::Float,
             SqlEngine::MySQL => DfType::Float,
         }
     }
